requirements.txt
flask
flask-cors
replicate
requestsflask
flask-cors
replicate
gunicorn
httpx
